requirements.txt
allianceauth
django-environ
gunicorn
aa-memberaudit
aa-memberaudit-securegroups
django-eveuniverse
aa-discordnotify
allianceauth-securegroups
allianceauth-corptools==2.5.5
gevent
django-health-check
django-debug-toolbar
